+++ R/estimateMissSBM.R
@@ -25,12 +25,14 @@
 #'  \item{"similarity": }{An R x R -> R function to compute similarities between node covariates. Default is
 #'         \code{missSBM:::l1_similarity}, that is, -abs(x-y). Only relevant when the covariates are node-centered
 #'         (i.e. \code{covariates} is a list of size-N vectors).}
-#'  \item{"threshold": }{V-EM algorithm stops stop when an optimization step changes the objective function
+#'  \item{"threshold": }{V-EM algorithm stops stop when an optimization step changes the objective function or the parameters
 #'         by less than threshold. Default is 1e-3.}
 #'  \item{"maxIter": }{V-EM algorithm stops when the number of iteration exceeds maxIter.
 #'        Default is 100 with no covariate, 50 otherwise.}
 #'  \item{"fixPointIter": }{number of fix-point iterations in the V-E step.
 #'        Default is 5 with no covariate, 2 otherwise.}
+#'  \item{"smoothing": }{character indicating what kind of smoothing should be use among "forward", "backward", "both" or "none". Default is "both".}
+#'  \item{"iterates": }{integer for the number of iterations of smoothing. Only relevant when smoothing is different from "none". Default is 1.}
 #'  \item{"cores": }{integer for number of cores used. Default is 1.}
 #'  \item{"trace": }{integer for verbosity (0, 1, 2). Default is 1. Useless when \code{cores} > 1}
 #' }
@@ -63,7 +65,7 @@
 #' theta <- list(mean = diag(.45,Q) + .05 ) # connectivity matrix
 #'
 #' ## Sampling parameters
-#' samplingParameters <- .5 # the sampling rate
+#' samplingParameters <- .75 # the sampling rate
 #' sampling  <- "dyad"      # the sampling design
 #'
 #' ## generate a undirected binary SBM with no covariate
@@ -84,8 +86,6 @@
 #' coef(myModel, "connectivity")
 #' predict(myModel)[1:5, 1:5]
 #'
-#' @import R6 parallel
-#' @import R6 parallel
 #' @export
 estimateMissSBM <- function(adjacencyMatrix, vBlocks, sampling, covariates = NULL, control = list()) {
 
@@ -99,8 +99,8 @@ estimateMissSBM <- function(adjacencyMatrix, vBlocks, sampling, covariates = NUL
   ## If nothing specified by the user, use covariates by default
   else if (is.null(control$useCovSBM)) control$useCovSBM <- TRUE
 
-  ## Defaut control parameters overwritten by user specification
-  ctrl <- list(threshold = 1e-3, trace = 1, cores = 1, imputation = "median", similarity = l1_similarity)
+  ## Default control parameters overwritten by user specification
+  ctrl <- list(threshold = 1e-3, trace = 1, cores = 1, imputation = "median", similarity = l1_similarity, iterates = 1, smoothing = "both")
   if (control$useCovSBM) {
     stopifnot(sampling %in% available_samplings_covariates)
     ctrl <- c(ctrl,list(maxIter = 50, fixPointIter = 2))
@@ -125,6 +125,9 @@ estimateMissSBM <- function(adjacencyMatrix, vBlocks, sampling, covariates = NUL
 
   ## Launch estimation of each missSBM_fit
   myCollection$estimate(ctrl)
+
+  ## Launch estimation of each missSBM_fit
+  myCollection$smooth(ctrl)
 
   ## Return the collection of adjusted missSBM_fit
   myCollection